Refactor TF tests for importer split, update pytorch tests.
