+++ tank/pytorch/wide_resnet50_2/wide_resnet50_2_test.py
@@ -1,33 +1,30 @@
 from shark.shark_inference import SharkInference
 from shark.shark_importer import SharkImporter
 from shark.iree_utils._common import check_device_drivers, device_driver_info
-from tank.model_utils import get_hf_model, compare_tensors
+from tank.model_utils import get_vision_model, compare_tensors
 from shark.parser import shark_args
 
 import torch
 import unittest
 import numpy as np
+import torchvision.models as models
 import pytest
 
 torch.manual_seed(0)
 
 
-class MiniLMModuleTester:
+class WideResnet50ModuleTester:
     def __init__(
         self,
-        dynamic=False,
-        device="cpu",
         save_mlir=False,
         save_vmfb=False,
     ):
-        self.dynamic = dynamic
-        self.device = device
         self.save_mlir = save_mlir
         self.save_vmfb = save_vmfb
 
-    def create_and_check_module(self):
-        model, input, act_out = get_hf_model(
-            "microsoft/MiniLM-L12-H384-uncased"
+    def create_and_check_module(self, dynamic, device):
+        model, input, act_out = get_vision_model(
+            models.wide_resnet50_2(pretrained=True)
         )
         shark_args.save_mlir = self.save_mlir
         shark_args.save_vmfb = self.save_vmfb
@@ -36,67 +33,65 @@ class MiniLMModuleTester:
             (input,),
             frontend="torch",
         )
-        minilm_mlir, func_name = mlir_importer.import_mlir(
-            is_dynamic=self.dynamic, tracing_required=True
+        mlir_module, func_name = mlir_importer.import_mlir(
+            is_dynamic=dynamic, tracing_required=False
         )
         shark_module = SharkInference(
-            minilm_mlir, func_name, device=self.device, mlir_dialect="linalg"
+            mlir_module, func_name, device=device, mlir_dialect="linalg"
         )
         shark_module.compile()
         results = shark_module.forward((input,))
         assert True == compare_tensors(act_out, results)
 
 
-class MiniLMModuleTest(unittest.TestCase):
+class WideResnet50ModuleTest(unittest.TestCase):
     @pytest.fixture(autouse=True)
     def configure(self, pytestconfig):
-        self.save_mlir = pytestconfig.getoption("save_mlir")
-        self.save_vmfb = pytestconfig.getoption("save_vmfb")
-
-    def setUp(self):
-        self.module_tester = MiniLMModuleTester(self)
+        self.module_tester = WideResnet50ModuleTester(self)
+        self.module_tester.save_mlir = pytestconfig.getoption("save_mlir")
+        self.module_tester.save_vmfb = pytestconfig.getoption("save_vmfb")
 
     def test_module_static_cpu(self):
-        self.module_tester.dynamic = False
-        self.module_tester.device = "cpu"
-        self.module_tester.create_and_check_module()
+        dynamic = False
+        device = "cpu"
+        self.module_tester.create_and_check_module(dynamic, device)
 
     def test_module_dynamic_cpu(self):
-        self.module_tester.dynamic = True
-        self.module_tester.device = "cpu"
-        self.module_tester.create_and_check_module()
+        dynamic = True
+        device = "cpu"
+        self.module_tester.create_and_check_module(dynamic, device)
 
     @pytest.mark.skipif(
         check_device_drivers("gpu"), reason=device_driver_info("gpu")
     )
     def test_module_static_gpu(self):
-        self.module_tester.dynamic = False
-        self.module_tester.device = "gpu"
-        self.module_tester.create_and_check_module()
+        dynamic = False
+        device = "gpu"
+        self.module_tester.create_and_check_module(dynamic, device)
 
     @pytest.mark.skipif(
         check_device_drivers("gpu"), reason=device_driver_info("gpu")
     )
     def test_module_dynamic_gpu(self):
-        self.module_tester.dynamic = True
-        self.module_tester.device = "gpu"
-        self.module_tester.create_and_check_module()
+        dynamic = True
+        device = "gpu"
+        self.module_tester.create_and_check_module(dynamic, device)
 
     @pytest.mark.skipif(
         check_device_drivers("vulkan"), reason=device_driver_info("vulkan")
     )
     def test_module_static_vulkan(self):
-        self.module_tester.dynamic = False
-        self.module_tester.device = "vulkan"
-        self.module_tester.create_and_check_module()
+        dynamic = False
+        device = "vulkan"
+        self.module_tester.create_and_check_module(dynamic, device)
 
     @pytest.mark.skipif(
         check_device_drivers("vulkan"), reason=device_driver_info("vulkan")
     )
     def test_module_dynamic_vulkan(self):
-        self.module_tester.dynamic = True
-        self.module_tester.device = "vulkan"
-        self.module_tester.create_and_check_module()
+        dynamic = True
+        device = "vulkan"
+        self.module_tester.create_and_check_module(dynamic, device)
 
 
 if __name__ == "__main__":
